Set up yamllint to lint YAML
It will be useful to check YAML streams because custom actions and
workflows are implemented as YAML documents and GitHub Actions, in
general, is based on YAML.

The command `poetry run -- yamllint .` is not preceded by a dollar sign
because it yields no output. The [documentation for the `markdownlint`
rule
`MD014`](https://github.com/updownpress/markdown-lint/blob/master/rules/014-commands-show-output.md)
explains that:

> [t]he dollar signs are unnecessary […] and should not be included.

Other commands are preceded by a dollar sign because they yield output:

> However, an exception is made when there is a need to distinguish
> between typed commands and command output, […].

This opinionated rule is based on [Ciro Santilli's Markdown style
guide](https://cirosantilli.com/markdown-style-guide/#dollar-signs-in-shell-code).